# Complete Rust rewrite with enhanced CI/CD pipeline and performance improvements