--- a/python/rez_tools/py.typed
+++ b/python/rez_tools/py.typed
@@ -1,0 +1,1 @@
+# Marker file for PEP 561