--- a/rez_tools/__init__.py
+++ b/rez_tools/__init__.py
@@ -1,7 +1,0 @@
-from pkg_resources import DistributionNotFound, get_distribution
-
-try:
-    __version__ = get_distribution(__name__).version
-except DistributionNotFound:
-    # Package is not installed.
-    __version__ = '0.0.0-dev'