--- a/rez_tools/reztoolsconfig.py
+++ b/rez_tools/reztoolsconfig.py
@@ -1,8 +1,0 @@
-import os
-
-# `rt` will search all tools from the below paths.
-tool_paths = [
-    os.path.normpath(os.path.expanduser("~/packages")),
-]
-
-extension = ".rt"